[SYNCOPE-1174] More robust conf params' default value management
This reverts commit 6634daaeebbe95fea9e1c104bc64d7b2e0d45e4b.

+++ core/provisioning-api/src/main/java/org/apache/syncope/core/provisioning/api/data/AccessTokenDataBinder.java
@@ -27,7 +27,7 @@ import org.apache.syncope.core.persistence.api.entity.AccessToken;
 
 public interface AccessTokenDataBinder {
 
-    Triple<String, String, Date> generateJWT(String subject, int duration, Map<String, Object> claims);
+    Triple<String, String, Date> generateJWT(String subject, long duration, Map<String, Object> claims);
 
     Pair<String, Date> create(String subject, Map<String, Object> claims, boolean replaceExisting);
